requirements.txt
flask
flask-mysqldb
pymysql
python-dotenv
